refactor(solver): split normalize order regression test
